+++ config.py
@@ -2,10 +2,10 @@
 FILES = {
     "search_log": "search_log.csv",
     "usage_counter": "usage_counter.txt",
-    "master_list": "saki_search/saki_master_list.csv",
-    "with_emails": "saki_search/saki_stores_with_emails.csv",
-    "without_emails": "saki_search/saki_stores_without_emails.csv",
-    "excluded_retailers": "saki_search/excluded_retailers.csv"
+    "master_list": "fenclaw_search/master_list.csv",
+    "with_emails": "fenclaw_search/stores_with_emails.csv",
+    "without_emails": "fenclaw_search/stores_without_email.csv",
+    "excluded_retailers": "fenclaw_search/excluded_retailers.csv"
 } 
 
 # FILES = {
