+++ fenclaw_search/excluded_retailers.csv
@@ -1,0 +1,41 @@
+retailer_name,reason
+Lush Cosmetics,major retail chain
+Sephora,major retail chain
+Bath & Body Works,major retail chain
+Bluemercury,major retail chain
+Aesop,major retail chain
+Ulta Beauty,major retail chain
+The Detox Market,major retail chain
+Chanel Fragrance and Beauty Boutique,major retail chain
+Charlotte Tilbury,major retail chain
+Glossier,major retail chain
+Planet Beauty,major retail chain
+The Cosmetics Company Store,major retail chain
+Benefit Cosmetics Boutique,major retail chain
+Shiseido Cosmetic Boutique,major retail chain
+Officina Profumo Farmaceutica di Santa Maria Novella,major retail chain 
+vons,major retail chain
+pavilions,major retail chain
+cvs,major retail chain
+walgreens,major retail chain
+walmart,major retail chain
+target,major retail chain
+bestbuy,major retail chain
+Barnes & Noble,major retail chain
+GameStop,major retail chain
+
+Kohl's,major retail chain
+Five Below,major retail chain
+2nd & Charles,major retail chain
+Books-A-Million,major retail chain
+OfficeMax,major retail chain
+Micro Center,major retail chain
+HobbyTown,major retail chain
+
+
+At Home,major retail chain
+Goodwill,major retail chain
+St Vincent de Paul,major retail chain
+REBEL Convenience Stores,major retail chain
+
+
