+++ optimal_radii.csv
@@ -46,3 +46,39 @@ Soap store,Los Angeles,33.85714923338188,-117.97759764792563,13000
 Soap store,Los Angeles,33.95996615406244,-118.06755217623692,13000
 Soap store,Los Angeles,33.7936048817827,-118.21310166048653,13000
 Soap store,Los Angeles,33.75433231270132,-118.06755217623692,13000
+tabletop gaming store,Charlotte,35.2271,-80.8431,50000
+fantasy gift shop,Charlotte,35.2271,-80.8431,50000
+fantasy gift shop,Raleigh,35.7796,-78.6382,50000
+witch store,Richmond,37.5407,-77.436,50000
+fantasy gift shop,Richmond,37.5407,-77.436,50000
+fantasy gift shop,Baltimore,39.2904,-76.6122,50000
+fantasy gift shop,Philadelphia,39.9526,-75.1652,50000
+fantasy gift shop,Detroit,42.3314,-83.0458,50000
+fantasy gift shop,Pittsburgh,40.4406,-79.9959,50000
+tabletop gaming store,Miami,25.7617,-80.1918,50000
+fantasy gift shop,Miami,25.7617,-80.1918,50000
+tabletop gaming store,Orlando,28.5383,-81.3792,50000
+fantasy gift shop,Orlando,28.5383,-81.3792,50000
+witch store,Memphis,35.1495,-90.049,50000
+fantasy gift shop,Memphis,35.1495,-90.049,50000
+witch store,Dallas,32.7767,-96.797,50000
+fantasy gift shop,Dallas,32.7767,-96.797,50000
+witch store,Tulsa,36.15398,-95.99277,50000
+fantasy gift shop,Tulsa,36.15398,-95.99277,50000
+witch store,Kansas City,39.0997,-94.5786,50000
+tabletop gaming store,Kansas City,39.0997,-94.5786,50000
+fantasy gift shop,Kansas City,39.0997,-94.5786,50000
+fantasy gift shop,Denver,39.7392,-104.9903,50000
+witch store,Phoenix,33.4484,-112.074,50000
+tabletop gaming store,Phoenix,33.4484,-112.074,50000
+fantasy gift shop,Phoenix,33.4484,-112.074,50000
+fantasy gift shop,Salt Lake City,40.7608,-111.891,50000
+witch store,Boise,43.615,-116.2023,50000
+tabletop gaming store,Boise,43.615,-116.2023,50000
+fantasy gift shop,Boise,43.615,-116.2023,50000
+witch store,Washington,38.9072,-77.0369,50000
+fantasy gift shop,Washington,38.9072,-77.0369,50000
+witch store,Las Vegas,36.1699,-115.1398,50000
+tabletop gaming store,Las Vegas,36.1699,-115.1398,50000
+fantasy gift shop,Las Vegas,36.1699,-115.1398,50000
+fantasy gift shop,Columbus,39.9612,-82.9988,50000
